 \name{gene.lists.ex}
 \docType{data}
 \alias{gene.lists.ex}
 \title{Eight gene lists}
 \description{
    This data contains artifical up- and down-regulated gene lists corresponding to two time points of two different experimental conditions. 
 }
 \value{
    Return a list of eight character vectors representing artifical up- and down-regulated gene lists.
 }
 \usage{data(gene.lists.ex)}
 \format{A list of eight character vectors.}
 \keyword{datasets}