 \name{result.kegg}
 \docType{data}
 \alias{result.kegg}
 \title{David search results}
 \description{
	This data contains an example of list of DAVIDFunctionalAnnotationChart objects computed by the DAVIDsearch function.
 }
 \value{
    Return a list of of DAVIDFunctionalAnnotationChart objects.
 }
 \usage{data(result.kegg)}
 \format{A list of of DAVIDFunctionalAnnotationChart objects.}
 \keyword{result.kegg}
          